extract AbstractClassAwareRector
