+++ src/Rector/AbstractRector.php
@@ -3,7 +3,6 @@
 namespace Rector\Rector;
 
 use PhpParser\Node;
-use PhpParser\Node\Stmt\Class_;
 use PhpParser\NodeTraverser;
 use PhpParser\NodeVisitorAbstract;
 use Rector\Contract\Deprecation\DeprecationInterface;
@@ -11,29 +10,6 @@ use Rector\Contract\Rector\RectorInterface;
 
 abstract class AbstractRector extends NodeVisitorAbstract implements DeprecationInterface, RectorInterface
 {
-    /**
-     * @var Class_|null
-     */
-    protected $classNode;
-
-    /**
-     * @param Node[] $nodes
-     * @return null|Node[]
-     */
-    public function beforeTraverse(array $nodes): ?array
-    {
-        $this->classNode = null;
-
-        foreach ($nodes as $node) {
-            if ($node instanceof Class_) {
-                $this->classNode = $node;
-                break;
-            }
-        }
-
-        return null;
-    }
-
     /**
      * @return null|int|Node
      */
@@ -48,28 +24,5 @@ abstract class AbstractRector extends NodeVisitorAbstract implements Deprecation
         }
 
         return null;
-    }
-
-    protected function getClassName(): string
-    {
-        if ($this->classNode === null) {
-            return '';
-        }
-
-        return $this->classNode->namespacedName->toString();
-    }
-
-    protected function getParentClassName(): string
-    {
-        if ($this->classNode === null) {
-            return '';
-        }
-
-        $parentClass = $this->classNode->extends;
-
-        /** @var Node\Name\FullyQualified $fqnParentClassName */
-        $fqnParentClassName = $parentClass->getAttribute('resolvedName');
-
-        return $fqnParentClassName->toString();
     }
 }
